+++ NEWBIE/TO-DO interactive-comments-section-main/JavaScript/card/contentUser.js
@@ -1,0 +1,36 @@
+import { createElement } from "../utilities-ui.js";
+
+export const createCommentHeader = () => {
+  const header = createElement("article", "content-title-card");
+  return header;
+};
+
+export const createUserAvatar = (user) => {
+  const contentUser = createElement("div", "content-user");
+
+  const photoUser = createElement("img");
+  photoUser.src = user.image.png;
+  contentUser.append(photoUser);
+
+  const nameUser = createElement("h2");
+  nameUser.textContent = user.username;
+  contentUser.append(nameUser);
+
+  return contentUser;
+};
+
+export const createButtonIcon = (svg, text) => {
+  const contentReply = createElement("div", "content-button-icon");
+
+  const iconReply = createElement("button", "icon-button");
+  iconReply.innerHTML = svg + text;
+  contentReply.append(iconReply);
+
+  return contentReply;
+};
+
+export const createText = (content) => {
+  const text = createElement("p", "text-comment");
+  text.textContent = content;
+  return text;
+};
