+++ NEWBIE/TO-DO interactive-comments-section-main/JavaScript/card/index.js
@@ -1,0 +1,2 @@
+export * from "./card";
+
